Refactoring of Kong Adapter
- Lots of changes, described in detail in DECISIONS.md
- Big refactoring of the entire code, making it cleaner and clearer
- Explicit support for oauth2-implicit (no pun intended)
- Better error handling
- Still missing: doPostImport

+++ kong/kong.js
@@ -2,9 +2,14 @@
 
 var async = require('async');
 var debug = require('debug')('kong-adapter:kong');
+var qs = require('querystring');
 var utils = require('./utils');
 
 var kong = function () { };
+
+// The maximum number of async I/O calls we fire off against
+// the Kong instance for one single call.
+const MAX_PARALLEL_CALLS = 10;
 
 kong.getKongApis = function (app, done) {
     debug('kong.getKongApis()');
@@ -60,6 +65,9 @@ function removeKongConsumerPlugins(apiList) {
 }
 
 kong.addKongApis = function (app, addList, done) {
+    // Bail out early if list empty
+    if (addList.length === 0)
+        return setTimeout(done, 0);
     debug('addKongApis()');
     // Each item in addList contains:
     // - portalApi: The portal's API definition, including plugins
@@ -88,6 +96,9 @@ kong.addKongApis = function (app, addList, done) {
 };
 
 kong.updateKongApis = function (app, sync, updateList, done) {
+    // Bail out early if list empty
+    if (updateList.length === 0)
+        return setTimeout(done, 0);
     debug('updateKongApis()');
     // Each item in updateList contains
     // - portalApi: The portal's API definition, including plugins
@@ -123,6 +134,9 @@ kong.updateKongApis = function (app, sync, updateList, done) {
 };
 
 kong.deleteKongApis = function (app, deleteList, done) {
+    // Bail out early if list empty
+    if (deleteList.length === 0)
+        return setTimeout(done, 0);
     debug('deleteKongApis()');
     // Each item in deleteList contains:
     // - kongApi
@@ -137,6 +151,9 @@ kong.deleteKongApis = function (app, deleteList, done) {
 };
 
 kong.addKongPlugins = function (app, addList, done) {
+    // Bail out early if list empty
+    if (addList.length === 0)
+        return setTimeout(done, 0);
     debug('addKongPlugins()');
     // Each entry in addList contains:
     // - portalApi: The portal's API definition
@@ -153,6 +170,9 @@ kong.addKongPlugins = function (app, addList, done) {
 };
 
 kong.updateKongPlugins = function (app, updateList, done) {
+    // Bail out early if list empty
+    if (updateList.length === 0)
+        return setTimeout(done, 0);
     debug('updateKongPlugins()');
     // Each entry in updateList contains:
     // - portalApi: The portal's API definition
@@ -170,6 +190,9 @@ kong.updateKongPlugins = function (app, updateList, done) {
 };
 
 kong.deleteKongPlugins = function (app, deleteList, done) {
+    // Bail out early if list empty
+    if (deleteList.length === 0)
+        return setTimeout(done, 0);
     debug('deleteKongPlugins()');
     // Each entry in deleteList contains:
     // - kongApi: Kong's API representation (for ids)
@@ -222,8 +245,22 @@ kong.deleteKongPlugins = function (app, deleteList, done) {
 ]
 */
 
-kong.getKongConsumers = function (app, done) {
+kong.getKongConsumers = function (app, portalConsumers, callback) {
     debug('getKongConsumers()');
+    async.mapLimit(
+        portalConsumers,
+        MAX_PARALLEL_CALLS, 
+        (portalConsumer, callback) => getKongConsumerInfo(app, portalConsumer, callback), 
+        function (err, results) {
+            if (err) {
+                console.error(err);
+                console.error(err.stack);
+                return callback(err);
+            }
+            debug('getKongConsumers() succeeded.');
+            return callback(null, results);
+        });
+    /*
     utils.kongGet(app, 'consumers?size=1000000', function (err, rawConsumerList) {
         if (err)
             return done(err);
@@ -238,7 +275,38 @@ kong.getKongConsumers = function (app, done) {
             return done(null, results);
         });
     });
+    */
 };
+
+function getKongConsumerInfo(app, portalConsumer, callback) {
+    debug('getKongConsumerInfo() for ' + portalConsumer.consumer.username);
+    async.waterfall([
+        callback => getKongConsumer(app, portalConsumer.consumer.username, callback),
+        (kongConsumer, callback) => enrichConsumerInfo(app, kongConsumer, callback) 
+    ], function (err, consumerInfo) {
+        if (err) {
+//            console.error(err);
+//            console.error(err.stack);
+            return callback(err);
+        }
+        // Note that the result may be null if the user is not present
+        return callback(null, consumerInfo);
+    });
+}
+
+function getKongConsumer(app, username, callback) {
+    debug('getKongConsumer(): ' + username);
+    utils.kongGet(app, 'consumers/' + username, function (err, consumer) {
+        if (err && err.status == 404) {
+            debug('getKongConsumer(): Not found.');
+            return callback(null, null); // Consider "normal", user not (yet) found
+        } else if (err) {
+            return callback(err);
+        }
+        debug('getKongConsumer(): Success, id=' + consumer.id);
+        return callback(null, consumer);
+    });
+}
 
 kong.CONSUMER_PLUGINS = [
     "acls",
@@ -250,6 +318,10 @@ kong.CONSUMER_PLUGINS = [
 
 function enrichConsumerInfo(app, kongConsumer, done) {
     debug('enrichConsumerInfo()');
+    if (!kongConsumer) {
+        debug('Not applicable, consumer not found.');
+        return done(null, null);
+    }
     var consumerInfo = {
         consumer: kongConsumer,
         plugins: {},
@@ -272,7 +344,7 @@ function enrichConsumerInfo(app, kongConsumer, done) {
 
 function enrichConsumerPlugins(app, consumerInfo, done) {
     debug('enrichConsumerPlugins()');
-    async.eachSeries(kong.CONSUMER_PLUGINS, function (pluginName, callback) {
+    async.each(kong.CONSUMER_PLUGINS, function (pluginName, callback) {
         utils.kongGet(app, 'consumers/' + consumerInfo.consumer.id + '/' + pluginName, function (err, pluginData) {
             if (err)
                 return callback(err);
@@ -311,7 +383,7 @@ kong.enrichConsumerApiPlugins = function (app, consumerInfo, /* optional */apiId
         // Do nothing then, no plugins
         return;
     }
-    utils.kongGet(app, 'apis/' + apiName + '/plugins?consumer_id=' + consumerId, function (err, apiPlugins) {
+    utils.kongGet(app, 'apis/' + apiName + '/plugins?consumer_id=' + qs.escape(consumerId), function (err, apiPlugins) {
         if (err) {
             // 404? If so, this may happen if the API was removed and there are still consumers on it. Ignore.
             if (err.status == 404)
@@ -341,6 +413,9 @@ function deleteKongConsumerApiPlugin(app, kongConsumer, kongApiPlugin, done) {
 }
 
 kong.addKongConsumerApiPlugins = function (app, addList, consumerId, done) {
+    // Bail out early if list empty
+    if (addList.length === 0)
+        return setTimeout(done, 0);
     debug('addKongConsumerApiPlugins()');
     async.eachSeries(addList, function (addItem, addCallback) {
         addKongConsumerApiPlugin(app, addItem.portalConsumer, consumerId, addItem.portalApiPlugin, addCallback);
@@ -352,6 +427,9 @@ kong.addKongConsumerApiPlugins = function (app, addList, consumerId, done) {
 };
 
 kong.patchKongConsumerApiPlugins = function (app, patchList, done) {
+    // Bail out early if list empty
+    if (patchList.length === 0)
+        return setTimeout(done, 0);
     debug('patchKongConsumerApiPlugins()');
     async.eachSeries(patchList, function (patchItem, patchCallback) {
         patchKongConsumerApiPlugin(app, patchItem.portalConsumer, patchItem.kongConsumer, patchItem.portalApiPlugin, patchItem.kongApiPlugin, patchCallback);
@@ -363,6 +441,9 @@ kong.patchKongConsumerApiPlugins = function (app, patchList, done) {
 };
 
 kong.deleteKongConsumerApiPlugins = function (app, deleteList, done) {
+    // Bail out early if list empty
+    if (deleteList.length === 0)
+        return setTimeout(done, 0);
     debug('deleteKongConsumerApiPlugins()');
     async.eachSeries(deleteList, function (deleteItem, deleteCallback) {
         deleteKongConsumerApiPlugin(app, deleteItem.kongConsumer, deleteItem.kongApiPlugin, deleteCallback);
@@ -376,10 +457,13 @@ kong.deleteKongConsumerApiPlugins = function (app, deleteList, done) {
 function patchKongConsumerApiPlugin(app, portalConsumer, kongConsumer, portalApiPlugin, kongApiPlugin, done) {
     debug('patchKongConsumerApiPlugin()');
     // Delete and re-add to make sure we don't have dangling properties
-    deleteKongConsumerApiPlugin(app, kongConsumer, kongApiPlugin, function (err) {
+    async.series([
+        callback => deleteKongConsumerApiPlugin(app, kongConsumer, kongApiPlugin, callback),
+        callback => addKongConsumerApiPlugin(app, portalConsumer, kongConsumer.consumer.id, portalApiPlugin, callback)     
+    ], function (err) {
         if (err)
             return done(err);
-        addKongConsumerApiPlugin(app, portalConsumer, kongConsumer.consumer.id, portalApiPlugin, done);
+        return done(null);
     });
 }
 
@@ -544,17 +628,49 @@ kong.updateKongConsumers = function (app, sync, updateList, done) {
     });
 };
 
-kong.deleteKongConsumers = function (app, deleteList, done) {
-    debug('deleteKongConsumers()');
-    // deleteItem has:
-    // - kongConsumer
-    async.eachSeries(deleteList, function (deleteItem, callback) {
-        utils.kongDelete(app, 'consumers/' + deleteItem.kongConsumer.consumer.id, callback);
-    }, function (err) {
+kong.deleteConsumerWithUsername = function (app, username, callback) {
+    debug('deleteConsumer() - username: ' + username);
+    utils.kongGet(app, 'consumers?username=' + qs.escape(username), function (err, consumerList) {
         if (err)
-            return done(err);
-        done(null);
+            return callback(err);
+        // Gracefully accept if already deleted
+        if (consumerList.total <= 0) {
+            console.error('Could not find user with username ' +  username + ', cannot delete');
+            return callback(null);
+        }
+        // This should be just one call, but the consumer is in an array, so this does not hurt.
+        async.map(consumerList.data, (consumer, callback) => deleteConsumerWithId(app, consumer.id, callback), function (err, results) {
+            if (err)
+                return callback(err);
+            callback(null);
+        });
     });
 };
+
+kong.deleteConsumerWithCustomId = function (app, customId, callback) {
+    debug('deleteConsumerWithCustomId() - custom_id: ' + customId);
+    utils.kongGet(app, 'consumers?custom_id=' + qs.escape(customId), function (err, consumerList) {
+        if (err)
+            return callback(err);
+        // Gracefully accept if already deleted
+        if (consumerList.total <= 0) {
+            console.error('Could not find user with custom_id ' + customId + ', cannot delete');
+            return callback(null);
+        }
+        if (consumerList.total > 1)
+            console.error('Multiple consumers with custom_id ' + customId + ' found, killing them all.');
+        // This should be just one call, but the consumer is in an array, so this does not hurt.
+        async.map(consumerList.data, (consumer, callback) => deleteConsumerWithId(app, consumer.id, callback), function (err, results) {
+            if (err)
+                return callback(err);
+            callback(null);
+        });
+    });
+};
+
+function deleteConsumerWithId(app, consumerId, callback) {
+    debug('deleteConsumerWithId(): ' + consumerId);
+    utils.kongDelete(app, 'consumers/' + consumerId, callback);
+}
 
 module.exports = kong;
